Start the clock from the oldest unconsumed build
Previously we were starting the clock as soon as a commit was pushed to
the branch our dependency's build came from.  That's not actionable,
since we actually need a build to be produced, and that build to be
pushed to the channel we're consuming before we *can* consume that
dependency.

This changes the calculation to "start the clock" for the SLA from the
time the oldest build currently published was produced.  This still
isn't ideal because a build could be produced but not published, sit
around for a week and then get published to the channel.  At that point
we would suddenly be out of SLA, even though there is nothing we could
have done.  Alas, the Maestro APIs don't seem to provide information
about *when* a particular build is published though, so this will have
to do for now.